Delta fix

+++ src/xrCore/FTimer.h
@@ -28,7 +28,7 @@ public:
 				CTimerBase		()		: qwStartTime(0),qwPausedTime(0),qwPauseAccum(0),bPause(FALSE)		{ }
 	ICF	void	Start			()		{	if(bPause) return;	qwStartTime = CPU::QPC()-qwPauseAccum;		}
 	ICF u64		GetElapsed_ticks()const	{	if(bPause) return	qwPausedTime; else return CPU::QPC()-qwStartTime-CPU::qpc_overhead-qwPauseAccum; }
-	IC	u32		GetElapsed_ms	()const	{	return u32(GetElapsed_ticks()*u64(1000)/CPU::qpc_freq );	}
+	IC	u64		GetElapsed_ms	()const	{	return u64(GetElapsed_ticks()*u64(1000)/CPU::qpc_freq );	}
 	IC	float	GetElapsed_sec	()const	{
 #ifndef _EDITOR
 		FPU::m64r	()			;
@@ -112,12 +112,12 @@ public:
 		return			(u32(GetElapsed_ticks()*u64(1000)/CPU::qpc_freq));
 	}
 	
-	IC	float			GetElapsed_sec	() const
+	IC	double			GetElapsed_sec	() const
 	{
 #ifndef _EDITOR
 		FPU::m64r		();
 #endif        
-		float			result = float(double(GetElapsed_ticks())/double(CPU::qpc_freq )	)	;
+		double			result = double(double(GetElapsed_ticks())/double(CPU::qpc_freq )	)	;
 #ifndef _EDITOR
 		FPU::m24r		();
 #endif
